+++ client/components/Navbar.jsx
@@ -17,9 +17,9 @@ export default function Navbar(){
 
     return (
         <ul className="Navbar">
-            <li onClick={home}>Home</li>
-            <li onClick={Login}>Login</li>
-            <li onClick={Profile}>Profile</li>
+            <li id='allbuttons' onClick={home}>Home</li>
+            <li id='allbuttons' onClick={Login}>Login</li>
+            <li id='allbuttons' onClick={Profile}>Profile</li>
         </ul>
     )
 }
