+++ client/styles/Homes.scss
@@ -1,0 +1,140 @@
+body {
+	background: linear-gradient( 86.3deg,  #97b7e1 3.6%, #79acef 87.6% );
+  font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
+	font-size: 20px;
+  background-size: 400% 400%;
+	animation: gradientbackground 4s ease-in infinite;
+	height: 100vh;
+}
+
+@keyframes gradientbackground {
+	0% {
+		background-position: 0% 50%;
+	}
+	50% {
+		background-position: 100% 50%;
+	}
+	100% {
+		background-position: 0% 50%;
+	}
+}
+
+.homepage{
+  animation: loadin 1.5s ease-in;
+}
+
+.home-top-title-button {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-around;
+}
+
+.input-container {
+  display: flex;
+  flex-direction: row;
+  margin-bottom: 20px;
+  justify-content: center;
+
+  .input-bar-home {
+    position: relative;
+    border-radius: 14px;
+    border: 1px solid;
+    background-color:  #A7BFE8;
+    width: 30%;
+    padding: 10px
+  }
+}
+
+.grid-container {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  gap: 20px;
+  overflow-y: auto;
+  max-height: 65vh;
+  justify-items: center;
+  
+
+  .box {
+    display: flex;
+    flex-direction: column;
+    justify-content: space-evenly;
+    align-items: center;
+    background-color: #6190E8;
+    padding: 50px;
+    width: 450px;
+    height: 320px;
+    word-wrap: break-word;
+    border-radius: 18px;
+    &:hover {
+      box-shadow: 5px 5px 10px black;
+      cursor:pointer
+    }
+    .api-image{
+      max-height: 200px;
+      max-width:200px;
+     }
+  }
+}
+
+#allbuttons{
+  width: 160px;
+  border-radius: 15px;
+  padding: 5px 5px;
+  border: none;
+  background-color: #A7BFE8;
+  &:hover{
+    cursor: pointer;
+    transform: scale(1.1);
+    color: white;
+  }
+}
+
+.newAddTech{
+  position:fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100vh;
+  animation: slow-load-in 1s ease-in;
+  background-color: rgb(0,0,0,0.2);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.formGroup{ width: 40%;
+min-width: 280px;
+max-width: 470px;
+position: relative;
+padding: 20px;
+background-color: #6190E8;
+border-radius: 16px;
+display: flex;
+flex-direction: column;
+justify-items: space-around;
+align-items: center;
+text-align: center;
+box-shadow: 10px 10px 5px rgba(0, 0, 0, 0.3);
+animation: loadin 1s ease-in;
+  .input-one{
+    border-radius: 6px;
+  border: none;
+  width: 80%;
+  padding: 8px;
+  margin-bottom: 4px;
+  background-color: #92b4ef;
+  color: black;
+  }
+}
+
+@keyframes loadin{
+  from{
+    opacity:0;
+    transform: translateX(-10x);
+  }
+  to{
+    opacity:1;
+    transform:translateX(0)
+  }
+}
